implement code-of-ethics section
