+++ app/guidelines/guidelines.tsx
@@ -1,0 +1,10 @@
+import CodeOfEthics from "./components/code-of-ethics"
+const Guidelines = () => {
+  return (
+    <div>
+      <CodeOfEthics/>
+    </div>
+  )
+}
+
+export default Guidelines
